rename list to gamelist
