+++ modules/ui/src/components/GameList/index.ts
@@ -1,0 +1,1 @@
+export * from "./GameList";
