--- modules/ui/src/components/List/index.ts
@@ -1,1 +1,0 @@
-export * from "./List";
